A new certificate comparison function
Comparing certificates for this service is slightly hard. We want to
verify that the certificate used in a request and the certificate being
presented as part of a +1 "I approve" signing request are the same.
However, when a signer signs a cert the Nonce, Signature and
SignatureKey fields are all overwritten with new values. A naive
comparison is therefore not possible.

Prior to this change the code would null out these fields and then do a
strict comparison.

The problem is that we were nulling out fields on the original
certificate request which meant that once it received its first +1 it
was now invalid and unable to be downloaded for a second +1. Lame.

Instead now we have a comparison function that knows to only compare the
values we know do not change (stated another way: the values we say are
not allowed to be changed).