Introduce PublicKeyFingerprint for requester
Prior to this commit, for some reason, all config files that dealt with
public keys did so by referencing fingerprints. For some reason I chose
to have the requester config use a path to a public key instead of a
fingerprint. This made ~zero sense.

With this change we now require only a fingerprint and this is the
preferred way of doing things. Documentation has been updated to show
the fingerprint option rather than the path option. The Path option
continues to work for backwards compatibility.