Support forcefully rejecting requests
Signers now have the option to "reject" a certificate request. Rejecting
a request permanently disables a cert request so that it will never be
signed. Signers can use this to reject certificate requests that violate
policy or are, perhaps, unreasonable. After being rejected no one has to
worry that the request gets signed at a later date.