Allow specifying a maximum certificate lifetime
A new signerd configuration option for limiting the maximum lifetime of
a requested certificate. This allows you to enforce policies that are,
for example, "never issue a certificate that is valid for more than 2
days."

By default there is no limit (max lifetime of 0).

This change is in preparation for creating a truly self-service portal.
One that can auto-sign certificates without requiring any approvals.