Make request id and serial number equal
Having a sequential serial number is silly and will make moving to HA
hard. The request id used to be 72 bits which fit really well in base32
encoding. The serial number is 64 bits on the cert so we make the
request id 64 bits.

Since 64 bits doesn't pack evenly in base32 the code now handles
improperly padded base32 encodings (i.e. missing the trailing ===).