Introducing the generate-config subcommand
Bootstrapping new users/requesters is harder than it should be due to
having to generating the configuration file on each host. This commit
introduces the generate-config subcommand that users can run to at least
bootstrap a configuration file.

Accompanying this is the /config/environments resource in sign_certd
that can return a list of configured environments on the server. This
enables generate-config to produce a pretty decent starting point (and
in some cases fully configured) configuration file for new users.